Add Paginator component to FindUsers
*Delete all alerts
