+++ src/Components/Find_Users/Find_Users.module.css
@@ -55,15 +55,7 @@
     padding-bottom: 10px;
     
 }
-.findUsers_pageNumber{
-    display: flex;
-    width: 800px;
-    flex-wrap: wrap;
-    gap: 3px;
-}
-.checkedPage{
-    font-weight: bold;
-}
+
